requirements.txt
streamlit==1.45.1 
pandas==2.1.4 
pyarrow==18.1.0 
google-api-python-client==2.149.0 
python-dotenv==1.0.1 
pyyaml==6.0.2 
plotly==5.24.1 
logging